add effect flag in Card.RegisterEffect  (#677)
* add enum effect_category

* add effect flag in Card.RegisterEffect

* fix effect_flag

* use uint64 flag

+++ libeffect.cpp
@@ -226,8 +226,8 @@ int32 scriptlib::effect_set_property(lua_State *L) {
 	check_param_count(L, 2);
 	check_param(L, PARAM_TYPE_EFFECT, 1);
 	effect* peffect = *(effect**) lua_touserdata(L, 1);
-	uint32 v1 = (uint32)lua_tointeger(L, 2);
-	uint32 v2 = (uint32)lua_tointeger(L, 3);
+	uint64 v1 = lua_tointeger(L, 2);
+	uint64 v2 = lua_tointeger(L, 3);
 	peffect->flag[0] = (peffect->flag[0] & INTERNAL_FLAGS) | (v1 & ~INTERNAL_FLAGS);
 	peffect->flag[1] = v2;
 	return 0;
